QueryLocations

SteamWebAPIs.get_countries — Function

get_countries()::Dict{String,Tuple{Bool,String}}

Summary: get_countries returns a Dict of Countries,key is country code, value is a Tuple, the first element of Tuple indicates whether there has stete, the second is countryname.

Example

julia> get_countries()
countries = Dict("ES" => (1, "Spain")...)

SteamWebAPIs.get_states — Function

get_states(countrycode::String)::Dict{String,String}

Summary: get_states returns a Dict of States by countrycode, key is state code, value is state name.

Arguments

countrycode: Country code to be queried.

Example

julia> get_states("CN")
Dict("24" => "Shanxi"...)

SteamWebAPIs.get_cities — Function

get_cities(countrycode::String,statecode::String)::Dict{Int,String}

Summary: get_cities returns a Dict of Cities by countrycode and statecode, key is city code, value is city name.

Arguments

countrycode: Country code to be queried.

statecode: State code to be queried.

Example

julia> get_cities("CN","01")
cites = Dict(10398 => "Wucheng"...)